picture_app/requirements.txt
beautifulsoup4==4.5.3
crcmod==1.7
Django==1.10.3
lxml==3.7.2
mysql-connector==2.1.4
MySQL-python==1.2.5
nltk==3.2.1
nose==1.3.7
numpy==1.11.2
oss2==2.2.2
Pillow==3.4.2
pygame==1.9.2
requests==2.12.3
South==1.0.2
vboxapi==1.0
